Create updated annotated ADS API grab
Andy and I sat down with Rahul's code and made it work.  Andy annotated the heck out of it, too.  Unfortunately, we realized we need to do more massaging of the results in the highlights field before we can scrape useful coords from it: that's the bit in red at the bottom.