Create working_code_14-04-2013
Look for the two variables you have to enter your information for (ADS API key and a file path for where you want the results1-10.txt files to end up).  More annotation, a bunch of debugging code that has been commented out, but there if you need to see what's happening every step of the way, fixed an indent error, and now it works!  Still need to update the regex with new logic.